codes/requirements.txt
openai
python-dotenv
json_repair
